+++ tests/test_readme.py
@@ -1,7 +1,5 @@
 import doctest
 
-import pytest
-
 
 # NOTE: We cannot test for the full table, as some Jobs depend on the time of execution
 #       e.g. a Job supposed to run on Weekday.MONDAY. The ordering between the Jobs scheduled