+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 
 ### API changes
 
-+ `Job`: `params` renamed to `kwargs` and introduced `argument` keyword analog to
++ `Job`: `params` renamed to `kwargs` and introduced `args` keyword analog to
   `sched` library
 
 ## 0.5.2